[feat] compatible with single channel diffusion
